requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
feedparser>=6.0.10
newspaper3k>=0.2.8
openai>=1.10.0
sqlalchemy>=2.0.25
apscheduler>=3.10.4
python-dotenv>=1.0.0
jinja2>=3.1.3
httpx>=0.26.0
lxml>=5.1.0
lxml_html_clean>=0.1.0
